backend/pyproject.toml
[project]
name = "urlsparkle"
version = "0.1.0"
dependencies = [
  "alembic>=1.14.1",
  "fastapi[standard]>=0.115.7",
  "psycopg[binary]>=3.2.4",
  "pydantic-settings>=2.7.1",
  "sentry-sdk[fastapi]>=2.20.0",
  "sqlmodel>=0.0.22",
  "validators>=0.34.0",
]
requires-python = ">=3.13"
authors = [{ name = "Patrick Godschalk", email = "patrick@kernelpanics.nl" }]
description = "URL shortening with FastAPI"
readme = "README.md"
license = { file = "LICENSE.txt" }

[dependency-groups]
dev = [
  "coverage>=7.6.10",
  "mypy>=1.14.1",
  "pre-commit>=4.1.0",
  "pytest>=8.3.4",
  "pytest-subtests>=0.14.1",
  "ruff>=0.9.3",
]

[tool.mypy]
strict = true
exclude = ["venv", ".venv", "alembic"]

[tool.pytest.ini_options]
# Add the current directory to Python's path
pythonpath = [".", "src"]

[tool.ruff]
target-version = "py313"
exclude = ["alembic"]

[tool.ruff.lint.pyupgrade]
# Preserve types, even if a file imports `from __future__ import annotations`.
keep-runtime-typing = true
